extracing button functions to parent
